Export reusable autoBelongsHook
