+++ packages/react/src/auto/polaris/inputs/relationships/PolarisAutoBelongsToForm.tsx
@@ -14,13 +14,10 @@ import {
   Text,
 } from "@shopify/polaris";
 import { MenuHorizontalIcon, PlusCircleIcon } from "@shopify/polaris-icons";
-import React, { useEffect, useState } from "react";
-import { useFormContext } from "../../../../useActionForm.js";
-import { get } from "../../../../utils.js";
+import React from "react";
+import { useHasBelongsForm } from "../../../../useHasBelongsForm.js";
 import { autoRelationshipForm } from "../../../AutoInput.js";
-import { RelationshipContext, useAutoRelationship, useRelationshipContext } from "../../../hooks/useAutoRelationship.js";
-import { useBelongsToController } from "../../../hooks/useBelongsToController.js";
-import { getRecordAsOption, useOptionLabelForField } from "../../../hooks/useRelatedModel.js";
+import { RelationshipContext } from "../../../hooks/useAutoRelationship.js";
 import type { OptionLabel } from "../../../interfaces/AutoRelationshipInputProps.js";
 import { RelatedModelOptionsPopover, RelatedModelOptionsSearch } from "./RelatedModelOptions.js";
 import { renderOptionLabel } from "./utils.js";
@@ -35,41 +32,27 @@ export const PolarisAutoBelongsToForm = autoRelationshipForm(
     secondaryLabel?: OptionLabel;
     tertiaryLabel?: OptionLabel;
   }) => {
-    const { field } = props;
-    const { path, metadata } = useAutoRelationship({ field });
     const {
-      setValue,
-      getValues,
-      formState: { defaultValues, submitCount, isSubmitSuccessful },
-    } = useFormContext();
-    const { record, relatedModelOptions } = useBelongsToController(props);
-    const [actionsOpen, setActionsOpen] = useState(false);
-    const [searchOpen, setSearchOpen] = useState(false);
-    const [modalOpen, setModalOpen] = useState(false);
-    const {
+      record,
+      actionsOpen,
+      modalOpen,
+      setActionsOpen,
+      searchOpen,
+      setSearchOpen,
+      setModalOpen,
       search,
       searchFilterOptions,
       pagination,
-      relatedModel: { records, fetching: isLoading },
-    } = relatedModelOptions;
-    const relationshipContext = useRelationshipContext();
-    const pathPrefix = relationshipContext?.transformPath ? relationshipContext.transformPath(props.field) : props.field;
-
-    const primaryLabel = useOptionLabelForField(field, props.primaryLabel);
-    const hasRecord = !!(record && !("_unlink" in record && record._unlink));
-    const recordOption = record ? getRecordAsOption(record, primaryLabel, props.secondaryLabel, props.tertiaryLabel) : null;
-
-    const defaultRecordId = get(defaultValues, path)?.id;
-
-    // if the child record is created we need to set the id to the default record id
-    // that comes from the response to the action mutation
-    useEffect(() => {
-      if (isSubmitSuccessful && record && !record.id && !("_link" in record) && !("_unlink" in record) && defaultRecordId) {
-        setValue(path + ".id", defaultRecordId);
-      }
-    }, [record, defaultRecordId, path, setValue, submitCount, isSubmitSuccessful]);
-
-    const parentName = metadata.name ?? "Unknown";
+      records,
+      isLoading,
+      pathPrefix,
+      hasRecord,
+      recordOption,
+      parentName,
+      path,
+      setValue,
+      getValues,
+    } = useHasBelongsForm(props);
 
     return (
       <>
